tested all endpoints

+++ app/repository.py
@@ -38,31 +38,38 @@ class AsyncpgItemRepository(AbstractItemRepository):
 
     async def create(self, item: ItemInDB, user_id: int) -> ItemInResponse:
         query = """
-            INSERT INTO items (title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, geom)
-            VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10::numeric, $11::numeric, ST_SetSRID(ST_MakePoint($11::numeric, $10::numeric), 4326))
-            RETURNING id, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
+            INSERT INTO items (category_id, offer_type, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, geom)
+            VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12::numeric, $13::numeric, ST_SetSRID(ST_MakePoint($13::numeric, $12::numeric), 4326))
+            RETURNING id, uuid, category_id, offer_type, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
         """
+        query2 = """INSERT INTO items_users (item_id, user_id) VALUES ($1, $2)"""
         async with self.conn as connection:
-            row = await connection.fetchrow(
-                query,
-                item.title,
-                item.description,
-                item.price,
-                item.currency,
-                item.amount,
-                item.measure,
-                item.terms_delivery,
-                item.country,
-                item.region,
-                item.latitude,
-                item.longitude,
-            )
-            item = ItemInResponse(**row)
-        return item
+            try:
+                row = await connection.fetchrow(
+                    query,
+                    item.category_id,
+                    item.offer_type,
+                    item.title,
+                    item.description,
+                    item.price,
+                    item.currency,
+                    item.amount,
+                    item.measure,
+                    item.terms_delivery,
+                    item.country,
+                    item.region,
+                    item.latitude,
+                    item.longitude,
+                )
+                item = ItemInResponse(**row)
+                await connection.execute(query2, item.id, user_id)
+                return item
+            except asyncpg.exceptions.ConnectionFailureError:
+                raise ValueError("Connection to the database failed")
 
     async def get_all(self, offset: int, limit: int) -> List[ItemInResponse]:
         query = """
-            SELECT id, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
+            SELECT id, uuid, category_id, offer_type, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
             FROM items
             ORDER BY id DESC
             OFFSET $1
@@ -74,7 +81,7 @@ class AsyncpgItemRepository(AbstractItemRepository):
 
     async def get_by_id(self, item_id: int) -> ItemInResponse:
         query = """
-            SELECT id, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
+            SELECT id, uuid, category_id, offer_type, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
             FROM items
             WHERE id = $1
         """
@@ -87,13 +94,15 @@ class AsyncpgItemRepository(AbstractItemRepository):
     ) -> ItemInResponse:
         query = """
             UPDATE items
-            SET title = $1, description = $2, price = $3, currency = $4, amount = $5, measure = $6, terms_delivery = $7, country = $8, region = $9, latitude = $10, longitude = $11
-            WHERE id = $12 AND id IN (SELECT item_id FROM items_users WHERE user_id = $13)
-            RETURNING id, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
+            SET category_id = $1, offer_type = $2, title = $3, description = $4, price = $5, currency = $6, amount = $7, measure = $8, terms_delivery = $9, country = $10, region = $11, latitude = $12, longitude = $13, geom = ST_SetSRID(ST_MakePoint($13::numeric, $12::numeric), 4326)
+            WHERE id = $14 AND id IN (SELECT item_id FROM items_users WHERE user_id = $15)
+            RETURNING id, uuid, category_id, offer_type, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
         """
         async with self.conn as connection:
             row = await connection.fetchrow(
                 query,
+                item.category_id,
+                item.offer_type,
                 item.title,
                 item.description,
                 item.price,
@@ -111,14 +120,8 @@ class AsyncpgItemRepository(AbstractItemRepository):
         return ItemInResponse(**row)
 
     async def delete(self, item_id: int, user_id: int) -> None:
-        query = """
-            DELETE FROM items_users
-            WHERE item_id = $1, user_id = $2
-        """
-        query2 = """
-            DELETE FROM items
-            WHERE id = $1
-        """
+        query = "DELETE FROM items_users WHERE item_id = $1 AND user_id = $2"
+        query2 = "DELETE FROM items WHERE id = $1"
         try:
             async with self.conn as connection:
                 await connection.execute(query, item_id, user_id)
@@ -128,7 +131,7 @@ class AsyncpgItemRepository(AbstractItemRepository):
 
     async def get_items_by_user_id(self, user_id: int) -> List[ItemInResponse]:
         query = """
-            SELECT i.id, i.title, i.description, i.price, i.currency, i.amount, i.measure, i.terms_delivery, i.country, i.region, i.latitude, i.longitude, i.created_at
+            SELECT i.id, i.uuid, i.category_id, i.offer_type, i.title, i.description, i.price, i.currency, i.amount, i.measure, i.terms_delivery, i.country, i.region, i.latitude, i.longitude, i.created_at
             FROM items i
             JOIN items_users iu ON i.id = iu.item_id
             WHERE iu.user_id = $1
@@ -141,7 +144,7 @@ class AsyncpgItemRepository(AbstractItemRepository):
         self, longitude: float, latitude: float, distance: int
     ) -> List[ItemInResponse]:
         query = """
-            SELECT id, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
+            SELECT id, uuid, category_id, offer_type, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
             FROM items
             WHERE ST_DWithin(geom, ST_SetSRID(ST_MakePoint($1, $2), 4326)::geography, $3)
         """
@@ -151,20 +154,20 @@ class AsyncpgItemRepository(AbstractItemRepository):
 
     async def get_filtered_items(
         self,
-        min_price: float,
-        max_price: float,
-        currency: str,
-        min_amount: int,
-        max_amount: int,
-        measure: str,
-        terms_delivery: str,
-        country: str,
-        region: str,
+        min_price: float = None,
+        max_price: float = None,
+        currency: str = None,
+        min_amount: int = None,
+        max_amount: int = None,
+        measure: str = None,
+        terms_delivery: str = None,
+        country: str = None,
+        region: str = None,
     ) -> List[ItemInResponse]:
         query = """
-            SELECT id, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
+            SELECT id, uuid, category_id, offer_type, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
             FROM items
-            WHERE (price >= $1 OR $1 IS NULL) AND (price <= $2 OR $2 IS NULL) AND (currency = $3 OR $3 IS NULL) AND (amount >= $4 OR $4 IS NULL) AND (amount <= $5 OR $5 IS NULL) AND (measure = $6 OR $6 IS NULL) AND (terms_delivery = $7 OR $7 IS NULL) AND (country = $8 OR $8 IS NULL) AND (region = $9 OR $9 IS NULL)
+            WHERE ($1::float IS NULL OR price >= $1) AND ($2::float IS NULL OR price <= $2) AND ($3::text IS NULL OR currency = $3) AND ($4::int IS NULL OR amount >= $4) AND ($5::int IS NULL OR amount <= $5) AND ($6::text IS NULL OR measure = $6) AND ($7::text IS NULL OR terms_delivery = $7) AND ($8::text IS NULL OR country = $8) AND ($9::text IS NULL OR region = $9)
         """
         async with self.conn as connection:
             rows = await connection.fetch(
@@ -180,39 +183,6 @@ class AsyncpgItemRepository(AbstractItemRepository):
                 region,
             )
         return [ItemInResponse(**row) for row in rows]
-
-    async def create_many(
-        self, items: List[ItemInDB], user_id: int
-    ) -> List[ItemInResponse]:
-        query = """
-            INSERT INTO items (title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, geom)
-            VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10::numeric, $11::numeric, ST_SetSRID(ST_MakePoint($11::numeric, $10::numeric), 4326))
-            RETURNING id, title, description, price, currency, amount, measure, terms_delivery, country, region, latitude, longitude, created_at
-        """
-        query2 = """
-            INSERT INTO items_users (item_id, user_id)
-            VALUES ($1, $2)
-        """
-        async with self.conn as connection:
-            items_list = []
-            for item in items:
-                row = await connection.fetchrow(
-                    query,
-                    item.title,
-                    item.description,
-                    item.price,
-                    item.currency,
-                    item.amount,
-                    item.measure,
-                    item.terms_delivery,
-                    item.country,
-                    item.region,
-                    item.latitude,
-                    item.longitude,
-                )
-                items_list.append(ItemInResponse(**row))
-                await connection.execute(query2, row["id"], user_id)
-        return items_list
 
 
 class AbstractUserRepository(ABC):
